refactor: update navigation and recording logic
- Change the home link in the NotFoundScreen to redirect to the new record screen.
- Update the TabLayout to include a new 'record' screen.
- Modify the FilesScreen to navigate to the index screen instead of the home screen.
- Replace the existing index screen with a redirect to the new record screen.
- Introduce a new RecordScreen component with audio recording functionality.

These changes enhance the navigation flow and introduce a dedicated screen for recording audio.
