+++ apps/playground/src/app/(tabs)/_layout.tsx
@@ -20,6 +20,12 @@ export default function TabLayout() {
             <Tabs.Screen
                 name="index"
                 options={{
+                    href: null,
+                }}
+            />
+            <Tabs.Screen
+                name="record"
+                options={{
                     title: 'Record',
                     tabBarLabel: ({ color, position }) => (
                         <Text
