+++ apps/playground/src/app/(tabs)/files.tsx
@@ -95,7 +95,7 @@ const FilesScreen = () => {
                     style={{ padding: 20 }}
                     buttonText="Record"
                     onButtonPress={() => {
-                        router.push('/')
+                        router.push('/(tabs)/index')
                     }}
                 />
             </ScreenWrapper>
